EBV netCDF properties class (S4)

EBV netCDF properties class (S4)

Value

S4 class containing the EBV netCDF properties

Slots

general

Named list. Elements: title, description, ebv_class, ebv_name, ebv_domain, references, source, project_name, project_url, creator_name, creator_institution, creator_email, contributor_name, publisher_name, publisher_institution, publisher_email, comment, keywords, id, history, licence, conventions, naming_authority, date_created, date_issued, entity_names, entity_type, entity_scope, entity_classification_name, entity_classification_url, taxonomy, taxonomy_lsid

spatial

Named list. Elements: wkt2, epsg, extent, resolution, crs_units, dimensions, scope, description

temporal

Named list. Elements: resolution, units, timesteps, dates

metric

Named list. Elements: name, description

scenario

Named list. Elements: name, description

ebv_cube

Named list. Elements: units, coverage_content_type, fillvalue, type

Note

If the properties class holds e.g. no scenario information this is indicated with an element called status in the list.
If you read an EBV netCDF based on an older standard, the properties will differ from the definition above. If the dataset does not encompass taxonomic info, the 'taxonomy' is NA. Besides, even if a dataset encompasses the taxonomy information, the 'taxonomy_lsid' can be NA.